line numbers can be usize

+++ src/chunk.rs
@@ -2,7 +2,7 @@ use crate::value::Value;
 
 #[derive(Debug)]
 pub struct Chunk {
-    pub code: Vec<(u8, u16)>,
+    pub code: Vec<(u8, usize)>,
     pub constants: Vec<Value>,
 }
 
@@ -14,13 +14,13 @@ impl Chunk {
         }
     }
 
-    pub fn write(&mut self, byte: u8, line: u16) {
+    pub fn write(&mut self, byte: u8, line: usize) {
         self.code.push((byte, line));
     }
 
-    pub fn add_constant(&mut self, value: Value) -> u8 {
+    pub fn add_constant(&mut self, value: Value) -> usize {
         self.constants.push(value);
-        self.constants.len() as u8 - 1
+        (self.constants.len() - 1) as usize
     }
 }
